requirements.txt
paddlepaddle>=2.3.1 
paddleocr==2.6.1.3 
vietocr>=0.3.8
pillow==9.5.0
torchvision==0.18.0